staking(component): clean component root
