+++ crates/core/app/src/app/mod.rs
@@ -26,7 +26,7 @@ use penumbra_sct::component::{StateReadExt as _, StateWriteExt as _};
 use penumbra_sct::epoch::Epoch;
 use penumbra_shielded_pool::component::{ShieldedPool, StateReadExt as _, StateWriteExt as _};
 use penumbra_stake::component::{
-    ConsensusUpdateRead, Staking, StateReadExt as _, StateWriteExt as _,
+    stake::ConsensusUpdateRead, Staking, StateReadExt as _, StateWriteExt as _,
 };
 use penumbra_transaction::Transaction;
 use prost::Message as _;
